refactor: updated file architecture
switched the file architecture to a full stack one with separete FE and BE directory to handle the frontend and backend respectively
