+++ src/frontend/assets/icons/index.ts
