+++ src/frontend/assets/images/index.ts
