+++ src/frontend/components/bottom-slider.ts
@@ -1,4 +1,4 @@
-import render from "../lib/dom";
+import render from "../../shared/lib/dom";
 
 export default function createBottomSlider() {
   const bottomSlider = document.createElement("div");
