+++ src/frontend/components/hadith-content-body.ts
