+++ src/frontend/components/hadith-content-footer.ts
@@ -1,8 +1,9 @@
-import render from "../lib/dom";
+import render from "../../shared/lib/dom";
 
 export default function createHadithContentFooter() {
   const hadithContentFooter = document.createElement("div");
-  hadithContentFooter.className = "w-full flex justify-between font-bold text-sm";
+  hadithContentFooter.className =
+    "w-full flex justify-between font-bold text-sm";
 
   const Left = FooterLeft();
   const Right = FooterRight();
@@ -14,7 +15,8 @@ export default function createHadithContentFooter() {
 
 function FooterLeft() {
   const hadithContentFooterLeft = document.createElement("div");
-  hadithContentFooterLeft.className = "w-full h-full flex flex-col justify-start text-left";
+  hadithContentFooterLeft.className =
+    "w-full h-full flex flex-col justify-start text-left";
   const line1 = document.createElement("p");
   const line2 = document.createElement("p");
   line1.textContent = "Narrated on the Authority of";
@@ -27,7 +29,8 @@ function FooterLeft() {
 
 function FooterRight() {
   const hadithContentFooterRight = document.createElement("div");
-  hadithContentFooterRight.className = "w-full h-full flex justify-end text-right";
+  hadithContentFooterRight.className =
+    "w-full h-full flex justify-end text-right";
   hadithContentFooterRight.textContent = "Hadith No: 736";
 
   return hadithContentFooterRight;
