+++ src/frontend/components/main-right-section.ts
@@ -1,5 +1,5 @@
 import { image3 } from "../assets/images";
-import render from "../lib/dom";
+import render from "../../shared/lib/dom";
 import createRightBottomPanel from "./right-bottom-panel";
 
 export default function createRightSection() {
