+++ src/shared/lib/dom.ts
