+++ src/shared/lib/result.ts
